Add betting support
Added some verbose reports to key functions. 
Created a copy of Request called RequestDebug that dumps the returned raw data. This is currently used when placing bets and querying the transaction list.
Created a new type CBfloat64 with it's own UnmarshalJSON. This type is used where citybet return parameters with quotes and without quotes depending on the value.
Add TransActionDetails() to get profit and loss for a market.
Add Transactions() to return the match status of bets on a market.
Add SubmitBet() that actually puts a bet on. Parameters dictate if it is a live or prereace bet.
Add SubmitBetRequest(), SubmitLiveBetRequest(), these call SubmitBet with the correct parameters to put a back bet on prerace and in play.
Add SubmitEat that puts eat bets on, called by SubmitEatRequest and SubmitLiveEatRequest to put eat bets on prerace or inplay.
ParseBid() returns the unmatched amounts of a bet response.
ParseTransacted() returns the matched state of a bet response.
DeleteBet() deletes bets, the parameters dictate if it is an indivual bet, all and back or eat.
DeletePendingBet() uses DeleteBet to delete a specific back bet.
DeletePendingEat() uses DeleteBet to delete a specific pending eat bet.
DeleteAllPendingBet() removes all pending back bets using DeleteBet.
DeleteAllPendingEat() deletes all pending eat bets using deleteBet().
Added Type CBfloat64.
Added TransactionDetail struct, and the parent structs including TransactionDetails.
Added Transaction struct.
Added BetResponse struct.