Duplicate race fix
Citibet randomly choose a market status to be the race status leading to races sometimes not going inplay when they do on the website.
As a fix they now return multiple copies of the race details, with an additional field racetype which denotes which tote pool/market this status applies to.
As a quick fix EventList now processes the list returned and sets a race's status to be inplay if any instance of that race is inplay.
Longer term this will be changed to return the full structure so the client can ultimately display the right race status depending on the market/pool selected.